Remove GirderBackend and sync_from_girder
This will no longer pull in Girder users upon login or registration.

Users will still be pushed back to Girder upon registration.
