+++ isic/login/girder.py
@@ -3,7 +3,6 @@ from functools import lru_cache
 import hashlib
 import random
 import string
-from typing import Optional
 
 from bson import ObjectId
 from django.conf import settings
@@ -16,25 +15,6 @@ from pymongo.database import Database
 def get_girder_db() -> Database:
     # Default database name is specified within ISIC_MONGO_URI
     return MongoClient(settings.ISIC_MONGO_URI).get_database()
-
-
-# Making this function private also makes it easier to mock in testing
-def _fetch_girder_user(query: dict) -> Optional[dict]:
-    girder_user = get_girder_db()['user'].find_one(query)
-
-    # coerce password to string
-    if girder_user and isinstance(girder_user['salt'], bytes):
-        girder_user['salt'] = girder_user['salt'].decode('utf-8')
-
-    return girder_user
-
-
-def fetch_girder_user_by_id(girder_user_id: str) -> Optional[dict]:
-    return _fetch_girder_user({'_id': ObjectId(girder_user_id)})
-
-
-def fetch_girder_user_by_email(email: str) -> Optional[dict]:
-    return _fetch_girder_user({'email': email.lower()})
 
 
 def create_girder_user(
