+++ isic/login/models.py
@@ -1,8 +1,5 @@
 import logging
-from typing import Optional
 
-from allauth.account.models import EmailAddress
-from django.conf import settings
 from django.contrib.auth.models import User
 from django.core.validators import RegexValidator
 from django.db import models
@@ -10,7 +7,6 @@ from django.db.models.signals import post_save
 from django.dispatch import receiver
 
 from isic.core.constants import MONGO_ID_REGEX
-from isic.login.girder import fetch_girder_user_by_email
 
 logger = logging.getLogger(__name__)
 
@@ -30,53 +26,8 @@ class Profile(models.Model):
     # check if the user password has changed.
     girder_salt = models.CharField(max_length=60, blank=True)
 
-    def sync_from_girder(self, girder_user: Optional[dict] = None) -> None:
-        if girder_user is None:
-            girder_user = fetch_girder_user_by_email(self.user.email)
-        if not girder_user:
-            # If this is a Django-native signup, the girder_user will not have been created yet
-            logger.info(f'Cannot retrieve girder_user for {self.user.email}.')
-            return
-
-        changed = False
-
-        if self.girder_id != str(girder_user['_id']):
-            self.girder_id = str(girder_user['_id'])
-            changed = True
-
-        if self.user.is_active != (girder_user.get('status', 'enabled') == 'enabled'):
-            self.user.is_active = girder_user.get('status', 'enabled') == 'enabled'
-            changed = True
-
-        if self.girder_salt != girder_user['salt']:
-            if isinstance(girder_user['salt'], bytes):
-                girder_user['salt'] = girder_user['salt'].decode('utf-8')
-            # An empty salt is stored as None in MongoDB, but should be '' here
-            self.girder_salt = girder_user['salt'] or ''
-            if not self.girder_salt:
-                self.user.set_unusable_password()
-            else:
-                self.user.password = f'bcrypt_girder${self.girder_salt}'
-            changed = True
-
-        if changed:
-            self.user.save()
-            self.save()
-
-        if girder_user['emailVerified'] is True:
-            EmailAddress.objects.update_or_create(
-                user=self.user,
-                email=self.user.email,
-                defaults={
-                    'primary': True,
-                    'verified': True,
-                },
-            )
-
 
 @receiver(post_save, sender=User)
 def create_or_save_user_profile(sender: type[User], instance: User, created: bool, **kwargs):
     if created:
-        profile = Profile.objects.create(user=instance)
-        if settings.ISIC_MONGO_URI:
-            profile.sync_from_girder()
+        Profile.objects.create(user=instance)
